refactor: notification service
